moved .vue into components folder
added transitions between router changes as it'll be useful to know how it's done later on.
