+++ src/components/Losers.vue
